libco: always use fiber API for co-routines on windows
The custom implementation of co-routines does not work well on Windows. Some API functions on Windows can throw an exceptions that have special meaning, used for debuggers for example. The OutputDebugString function throws the DBG_PRINTEXCEPTION_C which is basically just a signal for debugger to print a string. Under normal conditions this exception is automatically handled by Windows but this gets somehow broken once the code is executed under the co-routine, probably because of how the stack modification is implemented. Only x86 is affected because in x64 libco uses Fiber Windows API that works properly. I modified the libco to use always use fiber on windows.